v2.6.0（2020.11.22） 1. 重构[插入属性] a. 移除原来固定插入类型及简单的alloc init delloc，升级为关联已有类型 b. 移除随机拼接单词，使用智能名词替换 a. 新增‘百分比控制’开关 b. 新增‘Model后缀’开关，目的：避免Model归档或者数据转模型失败

+++ confuse_test/Confuse/oc/property/TestPropertyOther_VC.m
@@ -26,8 +26,8 @@
 - (void)viewDidLoad {
     [super viewDidLoad];
     _dataSource = [NSMutableArray array];
-    for (NSUInteger i = 0; i < 10; i++) {
-        [_dataSource addObject:[NSString stringWithFormat:@"test_%ld", i]];
+    for (int i = 0; i < 10; i++) {
+        [_dataSource addObject:[NSString stringWithFormat:@"test_%d", i]];
     }
     [_tableView reloadData];
 }
